requirements.txt
python-dotenv==1.0.0
openai==1.3.0
langchain==0.0.335
python-docx==1.0.0
spacy==3.7.2
pyyaml==6.0.1
